Move combigrid files to subfolders
