+++ combigrid/src/sgpp/combigrid/grid/IndexVectorIterator.hpp
@@ -6,8 +6,8 @@
 #pragma once
 
 #include <sgpp/globaldef.hpp>
-#include <sgpp/combigrid/FullGrid.hpp>
 #include <sgpp/combigrid/LevelIndexTypes.hpp>
+#include <sgpp/combigrid/grid/FullGrid.hpp>
 
 #include <iterator>
 #include <vector>
