+++ combigrid/src/sgpp/combigrid/operation/OperationEvalFullGrid.hpp
@@ -9,8 +9,9 @@
 #include <sgpp/base/datatypes/DataMatrix.hpp>
 #include <sgpp/base/datatypes/DataVector.hpp>
 #include <sgpp/base/operation/hash/OperationEval.hpp>
-#include <sgpp/combigrid/FullGrid.hpp>
-#include <sgpp/combigrid/IndexVectorRange.hpp>
+#include <sgpp/combigrid/LevelIndexTypes.hpp>
+#include <sgpp/combigrid/grid/FullGrid.hpp>
+#include <sgpp/combigrid/grid/IndexVectorRange.hpp>
 
 namespace sgpp {
 namespace combigrid {
