+++ combigrid/src/sgpp/combigrid/operation/OperationPole.hpp
