+++ combigrid/src/sgpp/combigrid/operation/OperationUPCombinationGrid.hpp
@@ -7,10 +7,10 @@
 
 #include <sgpp/globaldef.hpp>
 #include <sgpp/base/datatypes/DataVector.hpp>
-#include <sgpp/combigrid/CombinationGrid.hpp>
-#include <sgpp/combigrid/FullGrid.hpp>
-#include <sgpp/combigrid/OperationPole.hpp>
-#include <sgpp/combigrid/OperationUPFullGrid.hpp>
+#include <sgpp/combigrid/grid/CombinationGrid.hpp>
+#include <sgpp/combigrid/grid/FullGrid.hpp>
+#include <sgpp/combigrid/operation/OperationPole.hpp>
+#include <sgpp/combigrid/operation/OperationUPFullGrid.hpp>
 
 #include <vector>
